Added 'getting better/worse' to build messages' subject
git-svn-id: svn+ssh://rubyforge.org/var/svn/cerberus@130 65aa75ef-ce15-0410-bc34-cdc86d5f77e6

+++ test/test_helper.rb
@@ -58,9 +58,10 @@ class A#{rand(10000)}Test < Test::Unit::TestCase
 end"
     }
 
-    yield
-
-    FileUtils.rm test_case_name
+    if block_given?
+      yield
+      FileUtils.rm test_case_name
+    end
   end
 
   def add_application(app_name, url, options = {})
